backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
pytest>=8.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
openai>=1.30.0